refactor(dashboards): render row actions lazily
The popover passed its whole menu as antd's content prop, so every one of the
20 rows on a page built the full menu tree and instantiated two useMutation
hooks on every render, whether or not it was ever opened.

Move the body into ActionsPopoverContent behind a content render function plus
destroyTooltipOnHide, and extract the clone and lock-toggle mutations into
their own hooks. Only the row whose menu is open pays for it - which is also
what lets the per-dashboard permission checks be fired lazily next.

Note antd will not open a popover whose content is nullish, so the laziness has
to come from the render function rather than a conditional child.
